+++ tests/mod.rs
@@ -28,9 +28,10 @@ pub fn check_percentage_with_cli_args(minimum_coverage: f64, has_lines: bool, ar
              --root -r [DIR] 'directory'"
         ).get_matches_from(args);
 
-    let configs = ConfigWrapper::from(&matches).0;
+    let mut configs = ConfigWrapper::from(&matches).0;
     let mut res = TraceMap::new();
-    for config in &configs {
+    for config in configs.iter_mut() {
+        config.force_clean = false;
         let (t, _) = launch_tarpaulin(&config, &None).unwrap();
         res.merge(&t);
     }
@@ -59,6 +60,7 @@ pub fn check_percentage_with_config(
     env::set_current_dir(&test_dir).unwrap();
     config.manifest = test_dir;
     config.manifest.push("Cargo.toml");
+    config.force_clean = false;
 
     // Note to contributors. If an integration test fails, uncomment this to be able to see the
     // tarpaulin logs
@@ -82,7 +84,8 @@ pub fn check_percentage_with_config(
 }
 
 pub fn check_percentage(project_name: &str, minimum_coverage: f64, has_lines: bool) {
-    let config = Config::default();
+    let mut config = Config::default();
+    config.force_clean = false;
     check_percentage_with_config(project_name, minimum_coverage, has_lines, config);
 }
 
@@ -90,6 +93,7 @@ pub fn check_percentage(project_name: &str, minimum_coverage: f64, has_lines: bo
 fn incorrect_manifest_path() {
     let mut config = Config::default();
     config.manifest.push("__invalid_dir__");
+    config.force_clean = false;
     let launch = launch_tarpaulin(&config, &None);
     assert!(launch.is_err());
 }
@@ -98,6 +102,7 @@ fn incorrect_manifest_path() {
 fn proc_macro_link() {
     let mut config = Config::default();
     config.test_timeout = Duration::from_secs(60);
+    config.force_clean = false;
     let test_dir = get_test_path("proc_macro");
     config.manifest = test_dir.join("Cargo.toml");
     assert!(launch_tarpaulin(&config, &None).is_ok());
@@ -206,6 +211,7 @@ fn benchmark_coverage() {
     check_percentage(test, 0.0f64, true);
 
     let mut config = Config::default();
+    config.force_clean = false;
     config.run_types = vec![RunType::Benchmarks];
     check_percentage_with_config(test, 1.0f64, true, config);
 }
@@ -214,6 +220,7 @@ fn benchmark_coverage() {
 fn cargo_run_coverage() {
     let mut config = Config::default();
     config.command = Mode::Build;
+    config.force_clean = false;
     check_percentage_with_config("run_coverage", 1.0f64, true, config);
 }
 
@@ -224,6 +231,7 @@ fn examples_coverage() {
 
     let mut config = Config::default();
     config.run_types = vec![RunType::Examples];
+    config.force_clean = false;
     check_percentage_with_config(test, 1.0f64, true, config.clone());
 
     config.run_types.clear();
@@ -264,6 +272,7 @@ fn cargo_home_filtering() {
 
     let mut config = Config::default();
     config.test_timeout = Duration::from_secs(60);
+    config.force_clean = false;
     let restore_dir = env::current_dir().unwrap();
     let test_dir = get_test_path("HttptestAndReqwest");
     env::set_current_dir(&test_dir).unwrap();
@@ -290,6 +299,7 @@ fn rustflags_handling() {
     check_percentage("rustflags", 1.0f64, true);
     env::set_var("RUSTFLAGS", "--cfg=foo");
     let mut config = Config::default();
+    config.force_clean = false;
 
     let restore_dir = env::current_dir().unwrap();
     let test_dir = get_test_path("rustflags");
@@ -307,8 +317,10 @@ fn rustflags_handling() {
     assert_eq!(ret, 0);
 }
 
+#[test]
 fn follow_exes_down() {
     let mut config = Config::default();
     config.follow_exec = true;
+    config.force_clean = false;
     check_percentage_with_config("follow_exe", 1.0f64, true, config);
 }