+++ tests/workspaces.rs
@@ -11,6 +11,7 @@ fn package_exclude() {
     env::set_current_dir(&test_dir).unwrap();
     config.manifest = test_dir;
     config.manifest.push("Cargo.toml");
+    config.force_clean = false;
 
     config.all = true;
     let result = launch_tarpaulin(&config, &None);
@@ -40,6 +41,7 @@ fn package_exclude() {
 #[test]
 fn specify_package() {
     let mut config = Config::default();
+    config.force_clean = false;
     let test_dir = get_test_path("workspace");
     env::set_current_dir(&test_dir).unwrap();
     config.manifest = test_dir;