+++ CHANGELOG.md
@@ -11,6 +11,7 @@ compiler support. This isn't enabled so should have no effect it's just the
 start of the support work.
 - Now factors in rustflags from toml files #528
 - Now able to add to rustflags via CLI args and via tarpaulin config files
+- Added `--skip-clean` arg as an inverse to `--force-clean` and made cleaning default
 
 ### Changed
 - Make doctest prefix matching less specific as the naming convention changed again